+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.2.0 (unreleased)
+## 0.2.0 (2023-01-29)
 
 - Added `language` option
 - Changed `suggestions` method to filter by default